Move port from 5353 to 5301
While port 5353 is a nice DNS-y number, it's also the Multicast DNS, and the live madns server would occasionally get DNS-SD zeroconf networking requests sent to that port, trying to resolve the domain '_services._dns-sd._udp.local'.

I have no idea what the software on the other end is, but whatever it is, it seems to interpret madns's response of "not implemented" as "try again, immediately, until it is implemented", meaning my server would get hammered by dozens of requests per second until the thing on the other end finally gets the message.

So, the port is now 5301. I'm honestly surprised this didn't happen sooner.