Remove dropped feature from records list
This was on my list of records to add since the start, but now that I actually have to do it, I realise I didn't really know what I was getting at: it was originally going to be random Answer records with random qtypes and lengths and data, but if it generates a qtype where clients have to do string decompression, it'll just be an invalid record anyway. This isn't really testing for anything that can easily go wrong.